+++ frontend/app/global-not-found.tsx
@@ -10,8 +10,8 @@ import {
 } from "next/font/google";
 import type { Metadata } from "next";
 import styles from "@/app/not-found.module.css";
-import Navbar from "./(public)/_components/navbar";
-import Footer from "./(public)/_components/footer";
+import Navbar from "./_components/navbar";
+import Footer from "./_components/footer";
 import { getManagementSettings } from "./_lib/site-management";
 import darkenHex from "@/app/_utils/colorLightenDarken";
 import "@/app/global.css";
